reorganizes components by features and fixes headers
