+++ src/Components/Account/Account.js
@@ -1,6 +1,6 @@
 import { useState, useEffect } from "react";
 import { Link } from "react-router-dom";
-import { supabase } from "../supabaseClient";
+import { supabase } from "../../supabaseClient";
 
 const Account = ({ session }) => {
   const [loading, setLoading] = useState(true);
