+++ src/Components/Auth/AuthLogin.js
@@ -1,7 +1,7 @@
 import { useState } from "react";
 import { Link } from "react-router-dom";
 
-import { supabase } from "../supabaseClient";
+import { supabase } from "../../supabaseClient";
 
 export default function AuthLogin() {
   const [loading, setLoading] = useState(false);
