+++ src/Components/Auth/AuthSignup.js
@@ -1,6 +1,6 @@
 import { useState } from "react";
 
-import { supabase } from "../supabaseClient";
+import { supabase } from "../../supabaseClient";
 
 export default function AuthSignIn() {
   const [loading, setLoading] = useState(false);
